- fix(performance): Render blocking requests: move photoswipe to gallery.js
