+++ templates/pages/arcade.html.twig
@@ -2,6 +2,11 @@
 
 {% block title %}Arcade{% endblock %}
 
+{% block importmap %}
+    {{ parent() }}
+    {{ importmap('gallery') }}
+{% endblock %}
+
 {% block body %}
 <section class="bg-surface text-on-surface my-15">
     <div class="container">
