+++ templates/pages/photos.html.twig
@@ -2,6 +2,11 @@
 
 {% block title %}Photos{% endblock %}
 
+{% block importmap %}
+    {{ parent() }}
+    {{ importmap('gallery') }}
+{% endblock %}
+
 {% block body %}
 {% for type in photos_data.types %}
 <section class="{% if loop.index is even %}bg-surface-alt{% else %}bg-surface{% endif %} py-10">
