+++ templates/pages/projects.html.twig
@@ -2,6 +2,11 @@
 
 {% block title %}Projets{% endblock %}
 
+{% block importmap %}
+    {{ parent() }}
+    {{ importmap('gallery') }}
+{% endblock %}
+
 {% block body %}
 <section class="bg-surface text-on-surface">
     <div class="max-w-6xl mx-auto px-4 py-16 text-[14px] font-normal">
